imx: add initial Tufello support for imx platforms
This is basically a strip down version of libbt for the msm8992 mixed
with the BlueZ repo from CodeAurora which contains the Tufello code.

Also, it has been inspired from the broadcom libbt which is simple and
efficient.

Finally, the code has been cleaned up and indented unlike the code for
other platforms in this repo.

Signed-off-by: Gary Bisson <gary.bisson@boundarydevices.com>